ucd/ident: Separate features per property-group/feature
